### 2016-04-26 #40 @GShotwell
Convertr package: Extensive unit conversion with a shiny gadget. 
* 1. What does this package do? (explain in 50 words or less)
Provides extensive unit conversion in R, also includes RStudio add-in to assist the user in picking units.

* 2. Paste the full DESCRIPTION file inside a code block below.

```
Package: convertr
Title: Convert Between Units
Version: 0.1
Authors@R: person("Gordon", "Shotwell", email = "g.e.shotwell@gmaill.com",
                  role = c("aut", "cre"))
Description: Provides conversion functionality between a broad range of
    scientific, historical, and industrial unit types.
Imports: shiny(>=	0.13.2),
    miniUI(>= 0.1.1),
    DT(>= 0.1),
    rstudioapi(>=	0.5)
License: CC0

```

* 3. URL for the package (the development repository, not a stylized html page)
https://github.com/GShotwell/convertr
* 4. What data source(s) does it work with (if applicable)?
None
* 5. Who is the target audience?
Anyone with units to convert, mostly scientists and commodities analysts.
* 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?
`datamart` provides unit conversion, but with fewer units and a less extensible framework. (It's hard to add units to datamart). convertr has more units.
`udunits2` Provides the same or greater space of units as convertr, but requires access to an external API. Also udunits2 doesn't have an RStudio add-in. Convertr is based around a built-in conversion table so can be used without an API, it also has an RStudio add-in to assist the user in searching through available units.
* 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
    * [x ] This package does not violate the Terms of Service of any service it interacts with.
    * [x] The repository has continuous integration with Travis CI and/or another service
    * [ ] The package contains a vignette
    * [x ] The package contains a reasonably complete README with `devtools` install instructions
    * [x ] The package contains unit tests
    * [x ] The package only exports functions to the NAMESPACE that are intended for end users
* 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.

I agree, so long as I can figure out how to use Travis.

    * [ ] Are there any package dependencies not on CRAN?
No
Yes
    * [ ] Does the package have a CRAN accepted license?
Yes
    * [ ] Did `devtools::check()` produce any errors or warnings? If so paste them below.
No errors
* 9. Please add explanations below for any exceptions to the above:
* 10. If this is a resubmission following rejection, please explain the change in circumstances.
Not a resubmission
