### 2016-04-28 #40 @noamross
Thanks for the submission @GShotwell!  I must say we haven't developed solid criteria about overlap of packages yet.  I started a discussion over on the forum about it: https://discuss.ropensci.org/t/overlap-policy-for-package-onboarding/368.  I think we'll figure it out soon so we can move forward on this.  Please give us your thoughts!

In the meantime, Travis setup is easy and you can use `devtools::use_travis()` to start setup and you'll find more instructions here: https://docs.travis-ci.com/user/languages/r/
