### 2016-05-04 #40 @daattali
# Review

Overall, a great package that is implemented well and provides clear value.  I haven't used the other unit conversion packages, but I can see that `convertr` is very simple and easy to use and clearly has a very wide range of units available (way more than I knew existed!)

## General package guidelines/structure
- Perfect package name
- Good variable naming conventions and syntax
- You could consider adding a `CONDUCT.md` file; [more info](https://github.com/ropensci/packaging_guide#-code-of-conduct)
- Since you're using an MIT license, please add a license file; [more info](http://r-pkgs.had.co.nz/description.html#license)
- rOpenSci requires packages to have a `NEWS.md` file; [more info](https://github.com/ropensci/packaging_guide#-news)
- Installation and basic usage was smooth and fast with no errors on both Windows and Linux. Tests and R CMD check also ran fine for me locally :+1:
- "data_retrieval" folder: consider renaming this folder to "data-raw" as per [Hadley's guidelines](http://r-pkgs.had.co.nz/data.html)
- "data_retrieval" folder: there is a script named "data cleaning.R". It's recommended to not have spaces in file names, consider replacing the space with an underscore

## README
- In the first sentence, "cran" should be capitalized to "CRAN"
- In the second sentence, "a numerical vectors" should be "a numerical vector"
- Just a very personal opinion: the examples in the README could just show 1:5 instead of 1:20 so that the README isn't cluttered with so many numbers. 5 is enough to get the point
- The last example preceded by "/don't run" is missing a closing quote in the target unit
- I'm not sure I see the benefit of the "/don't run" in the README. Perhaps replacing that line with a comment saying the following is expected to result in an error is more clear
- After getting approved (or before? @noamross) you should add the rOpenSci footer image to the README; [more info](https://github.com/ropensci/packaging_guide#-readme)

## Tests
- There are no tests with vectors (all the tests only use atomic values)
- More tests would be a good idea, 3 positive and 3 negative isn't very comprehensive.  For example, there can be a test from unit X -> X, or a test that shows X -> Y == Y -> X
- None of the tests test what happens when a non-numeric is used as the first argument
- Similarly, it might be a good idea to test `NULL` or `NA` being passed as different arguments (not mandatory, just to be even more foolproof)

## Documentation
- Avoid exporting internal functions, consider using `#' @noRd` for internal functions (I can see conversion_table.Rd and is_supported_unit.Rd in the "man" folder); [more info](https://github.com/ropensci/packaging_guide#-documentation)
- The docs for `convert` say I can run `explore_units()` to see what units are available but that function doesn't exist
- Consider adding `@return` to the `convert` function documentation
- [Data description](https://github.com/GShotwell/convertr/blob/4209ff453145f751cd7d2710bf32370695f928f0/R/convertr.R#L16): when I load the dataframe, I get 1511 rows with 11 variables, not 1283x9 (`rp66_symbol` and `multi_unit` seem to be new variables)
- Typo: ["fromfrom"](https://github.com/GShotwell/convertr/blob/4209ff453145f751cd7d2710bf32370695f928f0/R/convertr.R#L8)
- Typo in roxygen formatting: ["#' Dictionary v2.2"](https://github.com/GShotwell/convertr/blob/4209ff453145f751cd7d2710bf32370695f928f0/R/convertr.R#L9) is in the middle of a line
- [The formula for conversion](https://github.com/GShotwell/convertr/blob/4209ff453145f751cd7d2710bf32370695f928f0/R/convertr.R#L14) has a mistake: based on looking at your code, the denominator is `C+D*X` rather than `C+B*X`
- I don't know much about that conversion formula - can you ever get a divide by 0 error with the formula? If so, try to catch it. If not, then never mind!
- The examples of the `convert` function could also benefit from having a `dontrun` example that shows incompatible types (the README shows this, so it'd be nice to have it in the function examples as well)

## Source code
- Generally good design decisions and coding style
- Potential for a helper function: [these lines](https://github.com/GShotwell/convertr/blob/4209ff453145f751cd7d2710bf32370695f928f0/R/convert.R#L33-L34) could benefit from a helper `get_record(unit)`.  It can also be used in the [gadget code](https://github.com/GShotwell/convertr/blob/4209ff453145f751cd7d2710bf32370695f928f0/R/gadget.R#L87-L90)
- The [`req` call](https://github.com/GShotwell/convertr/blob/4209ff453145f751cd7d2710bf32370695f928f0/R/gadget.R#L79) in the gadget needs to be namespaced to `shiny::req`
- In the gadget, the [code for creating the to/from dropdowns](https://github.com/GShotwell/convertr/blob/4209ff453145f751cd7d2710bf32370695f928f0/R/gadget.R#L50-L75) has some repetition and could be simplified by doing something like this (untested code, might not work out of the box):

        units_list <- reactive({
          if(input$si_unit == "All") {
            unique(conversion_table$catalog_symbol)
          } else {
            choices <- conversion_table[conversion_table$base_unit == input$si_unit, "catalog_symbol"]
            unique(choices)
          }
         })
        output$from_unit <- shiny::renderUI({ shiny::selectInput("from_unit", "From Unit", units_list())) })
        output$to_unit <- shiny::renderUI({ shiny::selectInput("to_unit", "To Unit", units_list())) })
Maybe I'm wrong, but I think something like that would work and it's cleaner and less repetitive
- It looks like you're only showing a maximum of 20 records in the table. If that's the case, consider removing the "Showing x out of x entries" (by adding `dom = ""` to the table's options list)

## Functionality
- In `convert`, when printing the names, consider using a different printing function instead of "print" so that the user won't see the ugly `[1] ""` and so that the user can have better control over it if it makes sense; [more info](https://github.com/ropensci/packaging_guide#-console-messages) (if this is purely only for debugging purposes then maybe it's not needed)
- Have you thought about the interaction of saving the results to a variable while printing the names? I think it's a bit weird: consider the command `foo <- convert(5, "km", "mi", print_names = TRUE)`. It will print the origin/target but not the result, and when you print out the variable, only the result will be printed. Have you thought about this and whether this is the desired behaviour?
- Even before seeing the mention of `explore_units()` I thought something like that would be a very useful addition to your package.  How do I know what all available units are? How do I know which conversions are valid? I could see great value in functions for both these questions.

## Addin/gadget
- The addin could benefit from better error checking and handling.  For example, if I type "1:10" into the numeric vector box, then while I type I see the ugly red Shiny error message and I get lots of errors in my console until the expression is complete. It would be much better user experience to ignore the expression until it is valid or at least catch the error from invalid expressions so that the console isn't flooded with errors and the user will see a more useful error message.
- Similarly, the input should be error checked. Since the gadget has an optional `vector` as an argument, it'd be a good idea to do some error checking on that input. Running `convertr:::convert_gadget("a")` or `convertr:::convert_gadget(letters)` does not do what I expected
- This is more of a personal taste thing, but I think the table looks awkward being restricted to 75% width instead of naturally filling up the entire row
- The UI looks a bit broken since the first two inputs are stuck to the left side without any whitespace, while the next two inputs have some padding. Looks quite weird and broken visually. You can fix this by wrapping all content inside any `fluidRow` in a `column()`.  [Before image](http://imgur.com/zPWjGWG) and [after image](http://imgur.com/Oqpfsc6)
- After choosing a "from" unit , it'd be helpful to only show the relevant "to unit" instead of all available units, where most of them will simply result in an error
- Assuming you end up implementing the functions mentioned earlier that list all available units and all available conversions from a specific unit, then the gadget could have another tab called "Help" or similar that will show ie. what all possible conversions are, the shortname --> fullname map of all units, or any other similarly useful info
- There seems to not be a difference between calling the gadget and running the addin. I feel like this is a bit technically wrong, because gadgets are meant to *return a value*, not to write to the editor. Addins do that. Renaming all the "gadget" occurrences to "addin" will make it more technically correct, but I think it'd be great to actually support both modes. If you run the app as a gadget, it should simply return a value (that is the intent of gadgets - to be a Shiny app that returns a value that can be stored to a variable), so that you can do something like `myvalue <- convertr::convert_gadget()`. If you run the addin, then it would run the gadget and in the end use the output value to write it to the document instead of storing it in a variable.  This is of course just a suggestion, but I think it would be a useful addition. You can see an example of this kind of functionality in the [colourPicker gadget/addin](https://github.com/daattali/shinyjs/blob/master/R/colourPickerGadget.R#L21-L25): calling `mycol <- shinyjs::colourPicker()` runs the gadget which means it will return a value, while running the addin will simply use that return value and write it like you are doing in your code. See [the very short colourPickerAddin function](https://github.com/daattali/shinyjs/blob/master/R/colourPickerGadget.R#L21-L25) to see how simple this can be

## Misc
- DESCRIPTION file: there's a tab character in the version number of rstudioapi and shiny
