### 2016-05-04 #40 @noamross
Thanks for your extensive review, @daattali!  Could you give an estimate of the hours you spent on it?

@GShotwell let @daattali or I know if you have questions.  We aim for responses back to review within two weeks.
