### 2016-05-05 #40 @GShotwell
Thanks so much for the amazing comments, I'll get right on them.
