### 2016-05-09 #40 @GShotwell
Thanks again for the great comments, I've learned a lot through this process. My comments are below. Mostly I made the recommended changes, but there were a couple of places where I decided not to, mostly in coding style areas (at what point to define a helper function).

** Tests
- I added all these tests

** Documentation
All implemented except:
- I removed the data documentation for the conversion table, I moved the table to a sysdata file within the R directory, so think the data documentation is a bit confusing. I added a convenence function to return the table if people are interested in taking a look at it.

** Source Code
- I opted not to create the `get_record()` function because it's slightly below my repetition limit, I think it's a bit harder to read with the helper rather than without.
- Same with the gadget, I think one more repetition and I'd create a new function, but think I'll leave it as is.

** Funcitonality
- I took out the print names, I think this function is covered by the add-in/gadget
- `explore_units()` is added back in

** Addin/gadget
- Error checking and messages are improved
- Style changes implemented
- the To unit now reacts to the from-unit selection
- I decided to not add the help menu, `explore_units()` has a lot of this functionality now, and I'm worried about making a gadget a bit too cluttered.
- I added a `return_value` argument to the gadget. So that you can either run it to generate an expression, or to do a conversion directly. The main reason I added the gadget along with the add-in was so that the user could start with a particular vector by calling `convert_gadget(vector)`. I can't really think of a good use case where it's helpful to return a converted value rather than the call which generates the value, but I think your right there might be one, so may as well provide an option to return a value.

** Misc
- All implemented
