### 2016-05-09 #40 @daattali
Looking much better! I think the new `explore_units()` is very useful.

I don't have too many new comments so I'll just dump them all together:

- In the README example, you have a type "This will produce and error" --> "This will produce an error"
- I would find it useful to mention in the README that the `explore_units()` function exists if you want to know what units are available
- Not a big deal, but the recommendation for the data folder is to be named "data-raw" rather than "data_raw"
- In convert.R, if you want to not run an example this is the correct syntax
    ```
    #' \dontrun{
    #' convert(1:20, "kg", "km2")
    #' }
    ```
- Also in that unrun example notice that there is a missing closing quotation mark
- In the documentation for explore_units() there doesn't seem to be a function title (the title is two sentences, I don't think that's what you meant?)
- Since the data table is now explicitly exported, in my opinion it would make sense to add some documentation about that table and say what each of the variables are. Previously it wasn't exported so it wasn't needed, but now that users are meant to see it, it should be documented
- Convert addin: Every time I select a new "from unit", I momentarily see an error in the gadget, and I see errors in the console. It's be awesome to fix that in order to have a nice smooth user experience without thinking that you did something wrong
