### 2016-05-10 #40 @GShotwell
Thanks! I've implemented all of these.
