### 2016-05-10 #40 @noamross
OK, we're about ready to go.  Before we transfer this to our repo, @GShotwell, I notice that there's a [build error](https://travis-ci.org/GShotwell/convertr/builds/129250060), which looks like `R CMD check` hanging on running the examples.  I believe that this is because functions launching shiny apps should be wrapped in `\dontrun{}` in examples.  Could you fix this? Then we should be good to go.
