### 2016-05-11 #40 @GShotwell
Oh drat, I did a check and then thought "hey, let's just add a couple more examples." I've fixed the error and convertr is now [passing](https://travis-ci.org/GShotwell/convertr/builds/129451926) the CMD check.
