### 2016-05-11 #40 @noamross
Great!  OK, next steps:

-   Add the footer to your README:

```
[![ropensci\_footer](http://ropensci.org/public_images/github_footer.png)](http://ropensci.org)
```
- Go to the Repo Settings --> Transfer Ownership and transfer to `ropenscilabs`!

Also, if you intend to submit to CRAN:
    -   I see you have one NOTE, that `data-raw` is a top-level directory. I believe you want to add this to .Rbuildignore. (Perhaps it used to be called `data_retrieval`?)
    -   I recommend using Travis to test against the development version of R.  To do so, change `sudo` to `false` in your `.travis.yml`, and adding the following.  This will run your tests across multiple R versions:

```
r:
  - release
  - devel
```
