### 2016-05-14 #40 @sckott
@GShotwell you should get an invitation to a team on `ropenscilabs` - then you should be able to do the transfer
