+++ NEWS.md
@@ -7,3 +7,4 @@ Version 0.1
 --------------------------------------------------------------------------------
 New Features
 * Added RStudio Add-in to build convertr expressions.
+ 
